==> 8_manuscript_markdown.Rmd <==
---
title: "Code used for paper titled *Individual and Census Tract-Level Correlates of Local SARS-CoV-2 Prevalence and Outcomes in Eastern Massachusetts*"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)



library(tidyverse)
library(naniar)
library(broom)
library(insight)
library(mgcv)
library(lubridate)
library(parameters)
library(aCRM)
library(performance)
library(missRanger)
library(pheatmap)
library(corrr)
library(car)
library(naniar)
library(ggrepel)
library(leaflet)
library(ggplot2)
library(DT)
library(sf)
library(performance)
library(postGIStools)
library(dbplyr)
library(DBI)
library(odbc)
library(RPostgreSQL)
library(effectsize)
library(lme4)
library(RODBC)
library(mapview)
library(parallel)
library(future)

library(metafor)

library(psych)

library(lctools)
#library(ape)

set.seed(9448)

round_df <- function(df, digits) {
  nums <- vapply(df, is.numeric, FUN.VALUE = logical(1))

  df[,nums] <- round(df[,nums], digits = digits)

  (df)
}

st_drop_geometry <- function(x) {
  if(inherits(x,"sf")) {
    x <- st_set_geometry(x, NULL)
    class(x) <- 'data.frame'
  }
  return(x)
}



extract_parameters <- function(i,df){
  parameters <- df[[i]]$result$parameters
  return(parameters)

}


extract_vif <- function(i,df){
  parameters <- df[[i]]$result$vif_df
  return(parameters)

}





number_of_cores <- detectCores()
plan(multisession, workers = number_of_cores)








patient_dim <- read_rds('/PHShome/cl331/covid_analysis/data/v2/patient_dim_v2.rds') %>% 
  mutate(PATIENT_NUM=as.character(PATIENT_NUM))  

race_dim <- patient_dim %>%
  dplyr::select(PATIENT_NUM,race_normalized, language_normalized, inscat1, race_normalized_only_race,reason_for_race_assignment) 


credentials <- list(dbname = '',
                    host="",
                    port=,
                    user = '',
                    password = '')







query <-
    "
  select a.shape_id,
  a.geoid,
  st_area(a.geographywkt)/1000000.0 as area_sq_km
  from exposome_pici.shapefile a
  where a.summarylevelid = ?summarylevelid
  and a.startdate = ?shapefile_startdate
  and a.enddate=?shapefile_enddate
  "




lab_regression_df_initial <- read_rds('/PHShome/cl331/covid_analysis/data/v2/covid_lab_pcr_aggregate.rds')  %>% 
  mutate(PATIENT_NUM=as.character(PATIENT_NUM))

patient_num_lab_df_covid_positive <- lab_regression_df_initial %>% 
  filter(any_pcr_pos == '1') %>% select(PATIENT_NUM) %>% unique()


inpatient_regression_df_initial <- read_rds('/PHShome/cl331/covid_analysis/data/v2/covid_inpatient.rds') %>% 
  mutate(PATIENT_NUM=as.character(PATIENT_NUM)) %>% 
  inner_join(patient_num_lab_df_covid_positive, by = 'PATIENT_NUM')



EMPI_address_geocoded <- read_rds('/PHShome/cl331/covid_analysis/data/v2/basic_info_address_subset.rds') %>% 
  mutate(PATIENT_NUM=as.character(PATIENT_NUM)) %>% 
  mutate(pop_dens = total_population/area_sq_km) %>%
  rename(age_test=AGE_IN_YEARS_NUM) %>%
  rename(Gender=SEX_CD) %>%
  mutate(non_white_pct = 1-white_pct) %>% dplyr::select(-race_normalized,-language_normalized,-inscat1) %>%
  left_join(race_dim, by='PATIENT_NUM') %>%
  mutate(pct_uninsured_dichotimized = if_else(pct_uninsured == 0.0,0,1)) %>%
  mutate(more_than_two_occupant_per_room_pct_dichotimized = if_else(more_than_two_occupant_per_room_pct == 0.0,0,1)) %>%
  mutate(race_normalized = if_else(is.na(race_normalized),'missing',race_normalized)) %>%
  mutate(language_normalized = if_else(is.na(language_normalized),'missing',language_normalized)) %>%
  mutate(inscat1 = if_else(is.na(inscat1),0,inscat1)) %>%
  mutate(race_normalized_only_race = if_else(is.na(race_normalized_only_race),'missing',race_normalized_only_race)) 


EMPI_address_geocoded$inscat1 <- factor(EMPI_address_geocoded$inscat1, levels = c("3", "0", "1","2"))
EMPI_address_geocoded$race_normalized <- factor(EMPI_address_geocoded$race_normalized, levels = c("white", "black", "hispanic","asian","other","missing"))


lab_regression_df <- lab_regression_df_initial %>% inner_join(EMPI_address_geocoded, by='PATIENT_NUM') %>%
  mutate_at(vars(total_population:units_in_structure_50_plus_pct), as.numeric) 


inpatient_regression_df <- inpatient_regression_df_initial %>% inner_join(EMPI_address_geocoded, by='PATIENT_NUM') %>%
  mutate_at(vars(total_population:units_in_structure_50_plus_pct), as.numeric) 


lab_regression_df <- lab_regression_df %>% 
  mutate(total_population = if_else(total_population <= 0.0, NA_real_,total_population)) %>%
  mutate(pop_dens = if_else(pop_dens <= 0.0, NA_real_,pop_dens))


inpatient_regression_df <- inpatient_regression_df %>% 
  mutate(total_population = if_else(total_population <= 0.0, NA_real_,total_population)) %>%
  mutate(pop_dens = if_else(pop_dens <= 0.0, NA_real_,pop_dens))

avg_population <- lab_regression_df %>% summarise(mean_population = mean(total_population, na.rm = TRUE), 
                                                  mean_pop_dens=mean(pop_dens, na.rm = TRUE))

lab_regression_df <- lab_regression_df %>% 
  mutate(total_population = if_else(is.na(total_population), avg_population$mean_population,total_population)) %>%
  mutate(pop_dens = if_else(is.na(pop_dens), avg_population$mean_pop_dens,pop_dens)) 




inpatient_regression_df <- inpatient_regression_df %>% 
  mutate(total_population = if_else(is.na(total_population), avg_population$mean_population,total_population)) %>%
  mutate(pop_dens = if_else(is.na(pop_dens), avg_population$mean_pop_dens,pop_dens)) 
  


per_capita_testing <- lab_regression_df %>% mutate(any_pcr_pos = as.numeric(any_pcr_pos)) %>%
  group_by(shape_id, total_population) %>% 
  tally() %>%  
  ungroup() %>% 
  mutate(per_capita_testing = n/total_population) %>% 
  select(-total_population,-n) 



lab_regression_df <- lab_regression_df %>% left_join(per_capita_testing, by='shape_id') %>%
  mutate(median_household_income = median_household_income / 1000.0) %>%
  mutate(median_home_value = median_home_value / 1000.0) %>%
  mutate(total_population_log = log10(total_population+1)) %>%
  mutate(units_in_structure_5_20_pct = units_in_structure_5_9_pct + units_in_structure_10_19_pct) %>%
  mutate(pop_dens_log = log10(pop_dens))



inpatient_regression_df <- inpatient_regression_df %>% left_join(per_capita_testing, by='shape_id') %>%
  mutate(median_household_income = median_household_income / 1000.0) %>%
  mutate(median_home_value = median_home_value / 1000.0) %>%
  mutate(total_population_log = log10(total_population+1)) %>%
  mutate(units_in_structure_5_20_pct = units_in_structure_5_9_pct + units_in_structure_10_19_pct) %>%
  mutate(pop_dens_log = log10(pop_dens))



lab_regression_df <- lab_regression_df %>% mutate(any_pcr_pos=as.numeric(any_pcr_pos))



lab_regression_df_pre_filter <- lab_regression_df
inpatient_regression_df_pre_filter <- inpatient_regression_df






lab_regression_df <- lab_regression_df %>% filter(VIP != 'E' & age_test >= 18) 

inpatient_regression_df <- inpatient_regression_df %>% filter(VIP != 'E' & age_test >= 18) 






environmental_variables <- c('over_65_all_pct',
                             'non_white_pct', 
                             'white_pct',
                             'black_pct',
                             'asian_pct',
                             'hispanic_pct',
                             'foreign_born_pct', 
                             'pct_private_insurance',
                             'pct_medicare_insurance',
                             'pct_medicaid_insurance',
                             'pct_uninsured',
                             'pct_uninsured_dichotimized',
                             'total_population_log',
                             'pop_dens_log',
                             'median_household_income',
                             'public_assistance_income',
                             'gini_index',
                             'poverty_under_100_pct', 
                             'poverty_under_150_pct',
                             'poverty_under_200_pct',
                             'nohs_pct', 
                             'college_pct', 
                             'unemployed_pct',
                             'essential_worker_pct', 
                             'more_than_one_occupant_per_room_pct',
                             'more_than_one_point_five_occupant_per_room_pct',
                             'more_than_two_occupant_per_room_pct',
                             'more_than_two_occupant_per_room_pct_dichotimized',
                             'household_2_plus_pct', 
                             'household_3_plus_pct',
                             'household_4_plus_pct',
                             'household_5_plus_pct',
                             'household_6_plus_pct',
                             'household_7_plus_pct',
                             'median_home_value',
                             'lack_plumbing_facilities_pct',
                             'units_in_structure_1_pct', 
                             'units_in_structure_2_plus_pct',
                             'units_in_structure_3_plus_pct',
                             'units_in_structure_5_plus_pct',
                             'units_in_structure_10_plus_pct',
                             'units_in_structure_20_plus_pct',
                             'units_in_structure_50_plus_pct',
                             'commute_walk_pct',
                             'commute_public_transportation_pct',
                             'commute_vehicle_pct',
                             'commute_work_from_home_pct',
                             'total_below_poverty',
                             'median_year_house_built',
                             'pop_dens')




















inpatient_regression_df <- inpatient_regression_df %>% mutate(inpatient_flag = as.numeric(inpatient_flag))



death_inpatient_df <- inpatient_regression_df %>% filter(inpatient_flag == '1') %>%
  filter(!is.na(VITAL_STATUS_CD)) %>%
  mutate(death_flag = if_else(VITAL_STATUS_CD == 'Y',1,0)) 




inpatient_census_tract <- inpatient_regression_df %>% 
  filter(inpatient_flag == '1') %>% 
  group_by(shape_id) %>% 
  tally() %>% 
  ungroup() %>%
  rename(inpatient_count = n)

death_census_tract <- death_inpatient_df %>% filter(death_flag == '1') %>% 
  group_by(shape_id) %>% 
  tally() %>%
  ungroup() %>%
  rename(death_count=n)
  

list_nums <- 1:length(environmental_variables)



con <- dbConnect(PostgreSQL(), dbname = "", user = "",
                 host = "",
                 password = "")


postgis_2013_2014_census_tract_sf <- get_postgis_query(con,
"select b.shape_id,
b.latitude,
b.longitude,
b.geoid,
b.statefip,
geometrywkt
from exposome_pici.shapefile b
where summarylevelid = '140'
and b.startdate = '2018-01-01'
and b.enddate = '2018-12-31' and statefip = ANY(ARRAY['25'])", geom_name = "geometrywkt") %>% st_as_sf(.) 


postgis_2013_2014_census_tract_sf <- st_set_crs(postgis_2013_2014_census_tract_sf, 4326) %>%
  mutate(shape_id=as.character(shape_id))


acs_variables <- lab_regression_df %>% 
  select(shape_id,total_population,white_pct,black_pct, asian_pct, hispanic_pct, 
         median_household_income, total_below_poverty, median_home_value, public_assistance_income, 
         gini_index, college_pct,nohs_pct, unemployed_pct, more_than_one_occupant_per_room_pct,
         more_than_one_point_five_occupant_per_room_pct, more_than_two_occupant_per_room_pct,
         pct_private_insurance, pct_medicare_insurance, pct_medicaid_insurance,over_65_all_pct,
         under_19_all_pct,poverty_under_100_pct,poverty_under_100_150_pct,poverty_under_150_200_pct,
         area_sq_km, commute_public_transportation_pct, median_year_house_built, lack_plumbing_facilities_pct,
         foreign_born_pct, essential_worker_pct, bluecollar_worker_pct, people_facing_worker_pct, household_3_plus_pct,
         household_4_plus_pct, household_5_plus_pct, household_6_plus_pct,household_7_plus_pct) %>% unique() 

acs_variables <- round_df(acs_variables, digits = 2)



num_positive_tests <- lab_regression_df %>% filter(any_pcr_pos == '1') %>% tally()

num_positive_tests <- num_positive_tests$n


patient_count_census_tract <- lab_regression_df %>% 
  group_by(shape_id, any_pcr_pos) %>%
  tally() %>% 
  pivot_wider(names_from = any_pcr_pos, values_from=n) %>% rename(negative=`0`,positive=`1`) %>%
  replace_na(list(positive=0,negative=0)) %>%
  mutate(total_patients=positive+negative) %>%
  mutate(pct_positive = positive/total_patients) %>%
  ungroup() %>%
  mutate(pct_positive = round(pct_positive,digits = 2))


spanish_census_tracts <- lab_regression_df %>% 
  filter( language_normalized == 'spanish') %>%
  group_by(shape_id, any_pcr_pos) %>% tally() %>%
    pivot_wider(names_from = any_pcr_pos, values_from=n) %>% rename(negative_spanish=`0`,positive_spanish=`1`) %>%
    replace_na(list(negative_spanish=0,positive_spanish=0))



hispanic_census_tracts <- lab_regression_df %>% 
  filter( race_normalized == 'hispanic') %>%
  group_by(shape_id, any_pcr_pos) %>% tally() %>%
    pivot_wider(names_from = any_pcr_pos, values_from=n) %>% 
  rename(negative_hispanic=`0`,positive_hispanic=`1`) %>%
    replace_na(list(negative_hispanic=0,positive_hispanic=0))


patient_count_census_tract <- patient_count_census_tract %>% 
  left_join(spanish_census_tracts,by='shape_id') %>%
    left_join(hispanic_census_tracts,by='shape_id') %>%
      replace_na(list(negative_spanish=0,positive_spanish=0)) %>%
      replace_na(list(negative_hispanic=0,positive_hispanic=0)) %>%
      mutate(pct_positive_all_tracts = positive/ num_positive_tests)



postgis_2013_2014_census_tract_sf_patient <- postgis_2013_2014_census_tract_sf %>%
  inner_join(patient_count_census_tract, by='shape_id') 

postgis_2013_2014_census_tract_sf_patient <- postgis_2013_2014_census_tract_sf_patient 

postgis_2013_2014_census_tract_sf_patient <- postgis_2013_2014_census_tract_sf_patient %>%
  mutate(count_category = case_when(pct_positive == 0.0 ~ '0%',
                                    pct_positive > 0.0 & pct_positive <= .10 ~ '0-10%',
                                    pct_positive > .10 & pct_positive <= .30 ~ '10-30%',
                                    pct_positive > .30 & pct_positive <= .50 ~ '30-50%',
                                    pct_positive > .50 & pct_positive < 1.0 ~ '50-100%',
                                    pct_positive == 1.0  ~ '100%'
                                    )) %>% left_join(acs_variables, by='shape_id') %>%
  mutate(count_category = factor(count_category, levels = c('0%','0-10%','10-30%','30-50%','50-100%','100%'))) %>%
  left_join(per_capita_testing, by='shape_id') %>%
  mutate(per_capita_testing = per_capita_testing * 1000) %>%
  left_join(inpatient_census_tract, by='shape_id') %>%
  left_join(death_census_tract, by='shape_id') %>%
  mutate(positive_per_1000=(positive * 1000)/total_population) %>%
  mutate(inpatient_per_1000=(inpatient_count*1000)/total_population) %>%
  mutate(death_per_1000=(death_count*1000)/total_population)


postgis_2013_2014_census_tract_sf_patient <- postgis_2013_2014_census_tract_sf_patient %>% filter(total_patients >=5)


```








This file contains code used to run the statistical analysis used in the paper titled *Individual and Census Tract-Level Correlates of Local SARS-CoV-2 Prevalence and Outcomes in Eastern Massachusetts*.  We present the summary statistics for the Census level variables used in this analysis as well as code for running the logistic mixed models for the three SARS-CoV-2 outcomes analyzed in the paper.


# Statistical distribution for census tract variables used in the analysis


Minimum, maximum, mean, median, 25th percentile, 75th percentile, 95th percentile, 99th percentile, and inter-quartile range for all Census tract variables used in this analysis.


```{r}



census_tract_info <- lab_regression_df %>% select(shape_id,environmental_variables) %>% unique() %>%
  left_join(postgis_2013_2014_census_tract_sf, by='shape_id') %>%
  mutate_at(vars(latitude,longitude),as.numeric) %>% filter(!is.na(latitude)) %>%
  filter(!is.na(longitude))

census_tract_info_df <- st_drop_geometry(census_tract_info) %>% dplyr::select(-shape_id,-latitude,-longitude,-statefip,-geometry)

census_tract_info_df_long <- census_tract_info_df %>% pivot_longer(-geoid)

#ggplot(census_tract_info_df_long, aes(x=value)) + geom_histogram() + facet_wrap(~name,scales = 'free')


census_tract_values_summary <- census_tract_info_df_long %>% group_by(name) %>% 
  summarise(min=min(value),
            max=max(value), 
            mean=mean(value), 
            median=quantile(value, probs=0.5, na.rm=TRUE),
            pct_25=quantile(value, probs=0.25, na.rm=TRUE),
            pct_75=quantile(value, probs=0.75, na.rm=TRUE),
            pct_95=quantile(value, probs=0.95, na.rm=TRUE),
            pct_99=quantile(value, probs=0.99, na.rm=TRUE),
            IQR = IQR(value)
            ) %>% 
  round_df(digits = 2)


DT::datatable(census_tract_values_summary)

```




# Lab Analysis

Code for analysis of risk of infection with SARS-CoV-2.  Risk of infection is based on positivity of viral polymerase chain reaction (PCR) testing for SARS-CoV-2 (lab test).

## Distribution of Age by Race and Primary Language

Age distribution of individuals with positive versus negative lab results for SARS-CoV-2.  Individuals also stratified by race and primary language spoken.

```{r}

ggplot(lab_regression_df, aes(x=age_test)) + 
  geom_histogram() + facet_wrap(~any_pcr_pos)


ggplot(lab_regression_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(any_pcr_pos~race_normalized)






ggplot(lab_regression_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(any_pcr_pos~language_normalized)




```

## Base Model 

Code used to run the base model for the SARS-CoV-2 lab test outcome.

```{r}
lab_matern_base <-  mgcv::gam(
  formula = any_pcr_pos ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    s(latitude, longitude, bs='gp', k=350, m=c(3,.01)),
  family = binomial(),
  data = lab_regression_df 
)


parameters_lab_matern_base <- parameters(lab_matern_base,exponentiate=TRUE)




saveRDS(parameters_lab_matern_base,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_lab_matern_base.rds')


```

### All Coefficients

Model statistics for all variables used in base model.

```{r}
DT::datatable(round_df(parameters_lab_matern_base, digits = 2))





```

### Coefficients by FDR < 0.05

Model statistics for all variables after filtering (FDR adjusted p-value < 0.05) in base model.

```{r}
DT::datatable(round_df(parameters_lab_matern_base %>% filter(p < 0.05), digits = 2))


```











## Base Plus Individual Environmental Variables

Code used to run the base-plus model for the SARS-CoV-2 lab test outcome.  This model iterates through all Census tract variables and fits a model with base variables plus the Census tract variable.


```{r}
gam_estimate_base_environmental_variable <- function(environmental_variable, df, formula_matern ){
  df <- df %>% rename(environmental_value=!!environmental_variable)
  
  model <- gam(formula = formula_matern, data = df,family = binomial())
  
   parameters_output <- parameters(model,exponentiate=TRUE)
   
   parameters_output$environmental_variable <- environmental_variable
   
#   parameters_output_stratified <- df %>% group_by(race_normalized) %>% 
  #   do(model_stratified=gam(formula = formula_matern_stratified, data = df,family = binomial())) %>%
  #   parameters(model,exponentiate=TRUE) %>%
  #   mutate(environmental_variable=environmental_variable) %>% ungroup()
   
  # parameters_output <- parameters_output %>% bind_rows(parameters_output_stratified)
   collinearity_matern_both <- check_collinearity(model, component='conditional')

      collinearity_matern_both$environmental_variable <- environmental_variable

   
   return(list(parameters=parameters_output,vif_df = collinearity_matern_both ))
  
}









gam_lab_matern_base_environment <- as.formula(any_pcr_pos ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized +
    inscat1 +
    per_capita_testing +
    environmental_value +
    s(latitude, longitude, bs='gp',k=350, m=c(3,.01))
    )




#gam_lab_parameters_base_environment_list <- furrr::future_map(.x=environmental_variables, 
#                                                 .f=safely(gam_estimate_base_environmental_variable),
#                                                 df=lab_regression_df,
#                                                 formula_matern=gam_lab_matern_base_environment)
  
  








#saveRDS(gam_lab_parameters_base_environment_list, '/PHShome/cl331/covid_analysis/data/v2/gam_lab_parameters_base_environment_list.rds')




gam_lab_parameters_base_environment_list <- readRDS('/PHShome/cl331/covid_analysis/data/v2/gam_lab_parameters_base_environment_list.rds')





list_nums <- 1:length(environmental_variables)







gam_lab_parameters_base_environment_df_subset <- map(.x=list_nums, .f=extract_parameters,df=gam_lab_parameters_base_environment_list) %>% 
  bind_rows() %>% filter(Parameter == 'environmental_value')


gam_lab_parameters_base_environment_df_subset_vif <- map(.x=list_nums, .f=extract_vif,df=gam_lab_parameters_base_environment_list) %>% 
  bind_rows()












coef_base_plus_rows <- gam_lab_parameters_base_environment_df_subset %>% 
  mutate(model='gam_base_plus') %>% 
  mutate(Parameter=environmental_variable)




```


### All Coefficients

Model statistics of all Census tract level variables after running all base-plus models.

```{r}
FDR_df_lab <- coef_base_plus_rows %>% 
  filter(!Parameter %in% c('age_test','race_normalizedblack','race_normalizedhispanic',
                           'race_normalizedasian','race_normalizedother','race_normalizedmissing',
                           'language_normalizedother','language_normalizedspanish','language_normalizedunknown',
                           'inscat10','inscat11','inscat12','GenderM','GenderU','per_capita_testing',
                           '(Intercept)'))




FDR_df_lab_fdr <- FDR_df_lab %>%     
  mutate(fdr=p.adjust(p, method="fdr")) %>%
  mutate(significance = if_else(fdr < 0.05,'FDR significant', 'FDR not significant'))

#DT::datatable(FDR_df_lab_fdr)

DT::datatable(round_df(FDR_df_lab_fdr, digits=4))





saveRDS(FDR_df_lab_fdr,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/FDR_df_lab_fdr.rds')



```

### Coefficients filtered by FDR < 0.05

Model statistics of all Census tract level variables (after filtering for coefficients with  FDR adjusted p-value < 0.05) after running all base-plus models.
```{r}
knitr::kable(round_df(FDR_df_lab_fdr %>% filter(significance == 'FDR significant'), digits = 2))


```



## Multivariate Analysis
Code used to run the multivariate analysis for SARS-CoV-2 lab test outcome.  This model selects all FDR significant Census tract variables from the base plus step and all base variables to fit a single multivariate model.


```{r}
lab_matern_fdr_filtered <-  mgcv::gam(
  formula = any_pcr_pos ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    black_pct +
    hispanic_pct + 
    foreign_born_pct +
    pct_private_insurance +
    pct_medicaid_insurance +
    pop_dens_log +
    poverty_under_150_pct +
    nohs_pct +
    college_pct+
    unemployed_pct +
    essential_worker_pct +
    household_5_plus_pct +
    commute_work_from_home_pct +
    s(latitude, longitude, bs='gp', k=350, m=c(3,.01)),
  family = binomial(),
  data = lab_regression_df 
)


parameters_lab_matern_fdr_filtered<- parameters(lab_matern_fdr_filtered,exponentiate=TRUE)




saveRDS(parameters_lab_matern_fdr_filtered,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_lab_matern_fdr_filtered.rds')



```

### All Coefficients

Model statistics of all variables used in multivariate model.


```{r}
DT::datatable(round_df(parameters_lab_matern_fdr_filtered, digits = 2))


```

### Coefficients by FDR < 0.05

Model statistics of all variables (after filtering for FDR adjusted p-value < 0.05) used in multivariate model.

```{r}
DT::datatable(round_df(parameters_lab_matern_fdr_filtered %>% filter(p < 0.05), digits = 2))


```










# Inpatient Analysis 

Code for analysis of hospitilization due to SARS-CoV-2 conditioned on receiving a positive SARS-CoV-2 lab test.

## Distribution of Race and Age

Age distribution of individuals hospitalized due to SARS-CoV-2 versus not hospitalized due to SARS-CoV-2 (among individuals with a positive SARS-CoV-2 lab test).  Individuals also stratified by race and primary language spoken.

```{r}



ggplot(inpatient_regression_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(inpatient_flag~race_normalized)






ggplot(inpatient_regression_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(inpatient_flag~language_normalized)



```








## Base Model 

Code used to run the base model for the hospitilization due to SARS-CoV-2.

```{r, warning=FALSE}

inpatient_matern_base <-  mgcv::gam(
  formula = inpatient_flag ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    s(latitude, longitude, bs='gp', k=350, m=c(3,.01)),
  family = binomial(),
  data = inpatient_regression_df 
)


parameters_inpatient_matern_base <- parameters(inpatient_matern_base,exponentiate=TRUE)





saveRDS(parameters_inpatient_matern_base,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_inpatient_matern_base.rds')


```

### All Coefficients

Model statistics for all variables used in base model.

```{r}
DT::datatable(round_df(parameters_inpatient_matern_base, digits = 2))



```

### Coefficients by FDR < 0.05

Model statistics for all variables after filtering (FDR adjusted p-value < 0.05) in base model.


```{r}
DT::datatable(round_df(parameters_inpatient_matern_base %>% filter(p < 0.05), digits = 2))


```













## Base Plus Individal Environmental Variables

Code used to run the base-plus model for hospitilization due to SARS-CoV-2 infection.  This model iterates through all Census tract variables and fits a model with base variables plus the Census tract variable.


```{r}




gam_inpatient_matern_base_environment <- as.formula(inpatient_flag ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized +
    inscat1 +
    per_capita_testing +
    environmental_value +
    s(latitude, longitude, bs='gp',k=350, m=c(3,.01))
    )





#gam_inpatient_parameters_base_environment_list <- furrr::future_map(.x=environmental_variables, 
#                                                 .f=safely(gam_estimate_base_environmental_variable),
#                                                 df=inpatient_regression_df,
#                                                 formula_matern=gam_inpatient_matern_base_environment)
  

#saveRDS(gam_inpatient_parameters_base_environment_list, '/PHShome/cl331/covid_analysis/data/v2/gam_inpatient_parameters_base_environment_list.rds')




gam_inpatient_parameters_base_environment_list <- readRDS('/PHShome/cl331/covid_analysis/data/v2/gam_inpatient_parameters_base_environment_list.rds')



gam_inpatient_parameters_base_environment_df_subset <- map(.x=list_nums, 
                                                     .f=extract_parameters,
                                                     df=gam_inpatient_parameters_base_environment_list) %>% 
  bind_rows() %>% filter(Parameter == 'environmental_value')


gam_inpatient_parameters_base_environment_df_subset_vif <- map(.x=list_nums, 
                                                               .f=extract_vif,
                                                               df=gam_inpatient_parameters_base_environment_list) %>% 
  bind_rows()



```






### All Coefficients

Model statistics of all Census tract level variables after running all base-plus models.


```{r}

FDR_df_inpatient <- gam_inpatient_parameters_base_environment_df_subset %>% 
  filter(!Parameter %in% c('age_test','race_normalizedblack','race_normalizedhispanic',
                           'race_normalizedasian','race_normalizedother','race_normalizedmissing',
                           'language_normalizedother','language_normalizedspanish','language_normalizedunknown',
                           'inscat10','inscat11','inscat12','GenderM','GenderU','per_capita_testing',
                           '(Intercept)'))




FDR_df_inpatient_fdr <- FDR_df_inpatient %>%   
  mutate(fdr=p.adjust(p, method="fdr")) %>%
  mutate(significance = if_else(fdr < 0.05,'FDR significant', 'FDR not significant')) %>%
  mutate(Parameter=environmental_variable)


DT::datatable(round_df(FDR_df_inpatient_fdr, digits = 2))





saveRDS(FDR_df_inpatient_fdr,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/FDR_df_inpatient_fdr.rds')


```



### Coefficients by FDR < 0.05

Model statistics of all Census tract level variables (after filtering for coefficients with  FDR adjusted p-value < 0.05) after running all base-plus models.



```{r}
knitr::kable(round_df(FDR_df_inpatient_fdr %>% filter(fdr < 0.05),digits=2))



```


## Multivariate GAM Model

Code used to run the multivariate analysis for hospitilization due to SARS-CoV-2 infection.  This model selects all FDR significant Census tract variables from the base plus step and all base variables to fit a single multivariate model.


```{r}

inpatient_matern_fdr_filtered <-  mgcv::gam(
  formula = inpatient_flag ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    hispanic_pct  +
    foreign_born_pct +
    pop_dens_log +
    median_household_income +
    poverty_under_200_pct +
    nohs_pct +
    college_pct +
    essential_worker_pct +
    more_than_two_occupant_per_room_pct_dichotimized +
    units_in_structure_2_plus_pct +
    s(latitude, longitude, bs='gp', k=350, m=c(3,.01)),
  family = binomial(),
  data = inpatient_regression_df 
)


parameters_inpatient_matern_fdr_filtered<- parameters(inpatient_matern_fdr_filtered,exponentiate=TRUE)




saveRDS(parameters_inpatient_matern_fdr_filtered,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_inpatient_matern_fdr_filtered.rds')


```

### All Coefficients

Model statistics of all variables used in multivariate model.


```{r}

DT::datatable(round_df(parameters_inpatient_matern_fdr_filtered, digits = 2))
```


### Coefficients by FDR < 0.05

Model statistics of all variables (after filtering for FDR adjusted p-value < 0.05) used in multivariate model.


```{r}
DT::datatable(round_df(parameters_inpatient_matern_fdr_filtered %>% filter(p < 0.05), digits = 2))
```




# Death Analysis


Code for analysis of death due to SARS-CoV-2 conditioned hospitilization due to SARS-CoV-2 lab infection.


## Distribution of Race and Age

Age distribution of individuals that have died due to SARS-CoV-2 versus not died due to SARS-CoV-2 (among individuals hospitalized due to SARS-CoV-2).  Individuals also stratified by race and primary language spoken.


```{r}


ggplot(death_inpatient_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(death_flag~race_normalized)






ggplot(death_inpatient_df, aes(x=age_test)) + 
  geom_histogram() + facet_grid(death_flag~language_normalized)



```









## Base Model 

Code used to run the base model for death due to SARS-CoV-2.

```{r, warning=FALSE}
death_matern_base <-  mgcv::gam(
  formula = death_flag ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    s(latitude, longitude, bs='gp',  k=100, m=c(3,0.00500)),
  family = binomial(),
  data = death_inpatient_df 
)


parameters_death_matern_base <- parameters(death_matern_base,exponentiate=TRUE)




saveRDS(parameters_death_matern_base,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_death_matern_base.rds')
```


### All Coefficients

Model statistics for all variables used in base model.

```{r}
DT::datatable(round_df(parameters_death_matern_base, digits = 2))


```

### Coefficients by FDR < 0.05

Model statistics for all variables after filtering (FDR adjusted p-value < 0.05) in base model.

```{r}
DT::datatable(round_df(parameters_death_matern_base %>% filter(p < 0.05), digits = 2))


```
















## Base Plus Individal Environmental Variables

Code used to run the base-plus model for death due to SARS-CoV-2 infection.  This model iterates through all Census tract variables and fits a model with base variables plus the Census tract variable.


```{r}

inpatient_regression_df %>% group_by(inpatient_flag,VITAL_STATUS_CD) %>% tally()




gam_death_matern_base_environment <- as.formula(death_flag ~ 
    age_test + 
    Gender +
    race_normalized + 
    language_normalized +
    inscat1 +
    per_capita_testing +
    environmental_value +
    s(latitude, longitude, bs='gp', k=100, m=c(3,0.00500))
    )




#gam_death_parameters_base_environment_list <- furrr::future_map(.x=environmental_variables, 
#                                                 .f=safely(gam_estimate_base_environmental_variable),
#                                                 df=death_inpatient_df,
#                                                 formula_matern=gam_death_matern_base_environment)
  
  

#saveRDS(gam_death_parameters_base_environment_list, '/PHShome/cl331/covid_analysis/data/v2/gam_death_parameters_base_environment_list.rds')




gam_death_parameters_base_environment_list <- readRDS('/PHShome/cl331/covid_analysis/data/v2/gam_death_parameters_base_environment_list.rds')


gam_death_parameters_base_environment_df_subset <- map(.x=list_nums, 
                                                     .f=extract_parameters,
                                                     df=gam_death_parameters_base_environment_list) %>% 
  bind_rows() %>% filter(Parameter == 'environmental_value')


gam_death_parameters_base_environment_df_subset_vif <- map(.x=list_nums, 
                                                               .f=extract_vif,
                                                               df=gam_death_parameters_base_environment_list) %>% 
  bind_rows()




```




```{r}

FDR_df_death<- gam_death_parameters_base_environment_df_subset %>% 
  filter(!Parameter %in% c('age_test','race_normalizedblack','race_normalizedhispanic',
                           'race_normalizedasian','race_normalizedother','race_normalizedmissing',
                           'language_normalizedother','language_normalizedspanish','language_normalizedunknown',
                           'inscat10','inscat11','inscat12','GenderM','GenderU','per_capita_testing',
                           '(Intercept)'))




FDR_df_death_fdr <- FDR_df_death %>%   
  mutate(fdr=p.adjust(p, method="fdr")) %>%
  mutate(significance = if_else(fdr < 0.05,'FDR significant', 'FDR not significant')) %>%
  mutate(Parameter=environmental_variable)




saveRDS(FDR_df_death_fdr,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/FDR_df_death_fdr.rds')


```

### All Coefficients

Model statistics of all Census tract level variables after running all base-plus models.


```{r}
DT::datatable(round_df(FDR_df_death_fdr, digits = 2))

```

### Coefficients by FDR < 0.05

Model statistics of all variables (after filtering for FDR adjusted p-value < 0.05) used in multivariate model.


```{r}

DT::datatable(round_df(FDR_df_death_fdr %>% filter(fdr < 0.05), digits = 2))



```


## Multivariate Analysis

Code used to run the multivariate analysis for death due to SARS-CoV-2 infection.  This model selects all FDR significant Census tract variables from the base plus step and all base variables to fit a single multivariate model.

```{r}

death_matern_fdr_filtered <-  mgcv::gam(
  formula = death_flag ~ 
    Gender +
    age_test + 
    race_normalized + 
    language_normalized + 
    inscat1 +
    per_capita_testing +
    s(latitude, longitude, bs='gp', k=100, m=c(3,0.00500)),
  family = binomial(),
  data = death_inpatient_df 
)


parameters_death_matern_fdr_filtered<- parameters(death_matern_fdr_filtered,exponentiate=TRUE)




saveRDS(parameters_death_matern_fdr_filtered,
        '/PHShome/cl331/covid_analysis/v3/covid_website/data/parameters_death_matern_fdr_filtered.rds')

```


### All Coefficients

Model statistics of all variables used in multivariate model.


```{r}


DT::datatable(round_df(parameters_death_matern_fdr_filtered, digits = 2))
```

### Coefficients by FDR < 0.05

Model statistics of all variables (after filtering for FDR adjusted p-value < 0.05) used in multivariate model.

```{r}

DT::datatable(round_df(parameters_death_matern_fdr_filtered %>% filter(p < 0.05), digits = 2))

```
